backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
praw==7.7.1
boto3==1.34.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cors==1.0.1
fastapi-cors==0.0.6